+++ maat-court-data-api/src/main/java/gov/uk/courtdata/hearing/crowncourt/impl/CrownCourtProcessHelper.java
@@ -32,7 +32,6 @@ public final class CrownCourtProcessHelper {
     }
 
     public String isBenchWarrantIssued(final HearingResulted hearingResulted) {
-
         List<String> offenceResultCodes = flattenResults(hearingResulted);
         return anyResultCodeMatch(xlatResultRepository.findByCjsResultCodeIn(), offenceResultCodes) ? YES : NO;
     }
